Move LocalRef to its own file.
