+++ cfg/Instructions.cc
@@ -221,56 +221,4 @@ string VariableUseSite::showRaw(const core::GlobalState &gs, const CFG &cfg, int
     }
 }
 
-int LocalRef::id() const {
-    return this->_id;
-}
-
-bool LocalRef::exists() const {
-    return this->_id > 0;
-}
-
-bool LocalRef::isAliasForGlobal(const core::GlobalState &gs, const CFG &cfg) const {
-    return this->data(cfg).isAliasForGlobal(gs);
-}
-
-bool LocalRef::isSyntheticTemporary(const core::GlobalState &gs, const CFG &cfg) const {
-    return this->data(cfg).isSyntheticTemporary(gs);
-}
-
-string LocalRef::toString(const core::GlobalState &gs, const CFG &cfg) const {
-    return this->data(cfg).toString(gs);
-}
-
-string LocalRef::showRaw(const core::GlobalState &gs, const CFG &cfg) const {
-    return this->data(cfg).showRaw(gs);
-}
-
-bool LocalRef::operator==(const LocalRef &rhs) const {
-    return this->_id == rhs._id;
-}
-
-bool LocalRef::operator!=(const LocalRef &rhs) const {
-    return this->_id != rhs._id;
-}
-
-LocalRef LocalRef::noVariable() {
-    return LocalRef(0);
-}
-
-LocalRef LocalRef::blockCall() {
-    return LocalRef(1);
-}
-
-LocalRef LocalRef::selfVariable() {
-    return LocalRef(2);
-}
-
-LocalRef LocalRef::unconditional() {
-    return LocalRef(3);
-}
-
-LocalRef LocalRef::finalReturn() {
-    return LocalRef(4);
-}
-
 } // namespace sorbet::cfg
