+++ cfg/Instructions.h
@@ -1,6 +1,7 @@
 #ifndef SORBET_INSTRUCTIONS_H
 #define SORBET_INSTRUCTIONS_H
 
+#include "cfg/LocalRef.h"
 #include "core/Context.h"
 #include "core/GlobalState.h"
 #include "core/LocalVariable.h"
@@ -11,41 +12,6 @@
 
 namespace sorbet::cfg {
 class CFG;
-
-class LocalRef final {
-    int _id;
-
-public:
-    LocalRef() : _id(0){};
-    LocalRef(int id) : _id(id){};
-    LocalRef(const LocalRef &) = default;
-    LocalRef(LocalRef &&) = default;
-    LocalRef &operator=(LocalRef &&) = default;
-    LocalRef &operator=(const LocalRef &) = default;
-
-    core::LocalVariable data(const CFG &cfg) const;
-    int id() const;
-    int minLoops(const CFG &cfg) const;
-    int maxLoopWrite(const CFG &cfg) const;
-    bool exists() const;
-    bool isAliasForGlobal(const core::GlobalState &gs, const CFG &cfg) const;
-    bool isSyntheticTemporary(const core::GlobalState &gs, const CFG &cfg) const;
-    std::string toString(const core::GlobalState &gs, const CFG &cfg) const;
-    std::string showRaw(const core::GlobalState &gs, const CFG &cfg) const;
-    bool operator==(const LocalRef &rhs) const;
-    bool operator!=(const LocalRef &rhs) const;
-
-    static LocalRef noVariable();
-    static LocalRef blockCall();
-    static LocalRef selfVariable();
-    static LocalRef unconditional();
-    static LocalRef finalReturn();
-};
-CheckSize(LocalRef, 4, 4);
-
-template <typename H> H AbslHashValue(H h, const LocalRef &m) {
-    return H::combine(std::move(h), m.id());
-}
 
 class VariableUseSite final {
 public:
